requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.1
requests==2.31.0
gunicorn==21.2.0
eventlet==0.33.3